+++ qGPSKet/models/plaquetteqGPS.py
@@ -21,7 +21,7 @@ class PlaquetteqGPS(nn.Module):
     M: int
     plaquettes: HashableArray
     dtype: DType = jnp.complex128
-    init_fun: NNInitFunc = normal(dtype=dtype)
+    init_fun: NNInitFunc = normal()
     to_indices: Callable = lambda samples : samples.astype(jnp.uint8)
     """
     syms is a tuple of two function representing the symmetry operations.
